Clean up linting warnings
