+++ service/history/service.go
@@ -37,7 +37,7 @@ type Service struct {
 	resource.Resource
 
 	status  int32
-	handler *handlerImpl
+	handler Handler
 	stopC   chan struct{}
 	params  *service.BootstrapParams
 	config  *config.Config
@@ -131,16 +131,10 @@ func (s *Service) Stop() {
 	s.GetMembershipMonitor().EvictSelf()
 
 	s.GetLogger().Info("ShutdownHandler: Waiting for others to discover I am unhealthy")
-	remainingTime = s.sleep(gossipPropagationDelay, remainingTime)
+	remainingTime = common.SleepWithMinDuration(gossipPropagationDelay, remainingTime)
 
-	s.GetLogger().Info("ShutdownHandler: Initiating shardController shutdown")
-	s.handler.controller.PrepareToStop()
-	s.GetLogger().Info("ShutdownHandler: Waiting for traffic to drain")
-	remainingTime = s.sleep(shardOwnershipTransferDelay, remainingTime)
-
-	s.GetLogger().Info("ShutdownHandler: No longer taking rpc requests")
-	s.handler.PrepareToStop()
-	remainingTime = s.sleep(gracePeriod, remainingTime)
+	remainingTime = s.handler.PrepareToStop(remainingTime)
+	remainingTime = common.SleepWithMinDuration(gracePeriod, remainingTime)
 
 	close(s.stopC)
 
@@ -148,14 +142,4 @@ func (s *Service) Stop() {
 	s.Resource.Stop()
 
 	s.GetLogger().Info("history stopped")
-}
-
-// sleep sleeps for the minimum of desired and available duration
-// returns the remaining available time duration
-func (s *Service) sleep(desired time.Duration, available time.Duration) time.Duration {
-	d := common.MinDuration(desired, available)
-	if d > 0 {
-		time.Sleep(d)
-	}
-	return available - d
 }
